chore: Updates supported browsers
Configurations and settings in package.json and tsconfig.base.json have been updated to support newer versions of browsers in production as specified in the "browserslist". The 'react-cookie-service' package had its type updated to "module" as well as an enhancement to its production browser support. These changes ensure better compatibility and performance in newer browsers, while also preparing for future upgrades of ECMAScript and maintaining alignment with our dependencies.