chore: "Update browser compatibility in package.json files"
Increased the range of browser versions for compatibility in main package.json file to ensure the application runs smoothly and is up to date on older yet still widely used versions of the listed browsers. Removed unnecessary 'browserslist' from 'react-cookie-service' package.json, as this detail should be defined at the main project level, not within individual packages. This helps to avoid inconsistencies and potential conflicts.